Updated version number to 2.5-alpha2 Enable all admins to view logs Support for case sensitive operators for dynamic queries Allow use of decimals when specifying y axis range

+++ art/src/art/WEB-INF/classes/art/graph/ExportGraph.java
@@ -224,8 +224,8 @@ public class ExportGraph {
 		boolean showTooltips = false;
 		boolean showUrls = false;
 		String bgColor;
-		int from; //y axis range minimum
-		int to; //y axis range maximum
+		double from; //y axis range minimum
+		double to; //y axis range maximum
 
 		ChartTheme currentChartTheme = null;
 		BarPainter currentBarPainter = null;
